// marketplace-server/src/types/file-types.ts
export type ImageExtension =
  | 'jpg'
  | 'jpeg'
  | 'png'
  | 'webp'
  | 'tiff'
  | 'gif'
  | 'heif'
  | 'avif'
  | 'jp2';